FluidSimulationSetup
First Upload of the Fluid Simulation Project containing a collection of physics-based simulation scripts designed to model particle interactions, gravitational forces, collisions, and fluid dynamics. The project aims to optimize CPU/GPU performance while maintaining accuracy and efficiency in simulations.

Additionally, this repository includes a downloadable premium asset: "Shapes", a digital asset designed to enhance the visualization and functionality of simulations.

Personal Scipts:
particleScripts: main driver of the simulation using immediate mode so simulate frame by frame

particleArrayScript: sets up most of the particle logic including properties like position and velocity for each particle object in an array.

Gravity: a function that multiplies a particles velocity by fixed delta time to get the new position as well as a dampening float variable

collisions: slightly misdirected as this is the box that the particles are incased in and only has a "collision" towards this box

FluidDynamics: density function; This calculates the density of all particles inside of an influence radius based off all or a given query of particles. pressure function; calculates and applies a pressure force to a particle by incrementing the velocity based on all or query particles